+++ test/test_runtime_contract.rb
@@ -71,6 +71,7 @@ plasmoid = project_root.join("kde/plasmoids/org.xnix.compatibilitycenter/metadat
   GetCompatibilityActionQueue
   GetCompatibilityActionReviewReceipt
   GetCompatibilityCenterSummary
+  GetKDECenterPage
 ].each do |method|
   assert(contract.include?("name=\"#{method}\""), "D-Bus contract must expose #{method}")
 end